suites/requirements.txt
sh
path.py
PyYAML==3.10
nose
xmltodict
requests